The Simulation of Car Wash Queuing Model
The purpose of this project is to simulate car wash queuing model which
allows people (or car wash station’s owner) to keep track of the system
state so they can see the throughput, response time, average response
time, average waiting time, and server utilization.

The significance of this queuing model breaks down to three components.
First is arrivals (or inputs), second is queue discipline (or waiting in
line), and last is the service facility. Based on above three
components, we have system of Car Wash, customers are cars, and the
server is the car wash facility.

In this car wash model, it can have any number of cars arriving, any
number of cars waiting in queue, and number of cars being washed
depending on whether if it is single-server or multi-server system.
Therefore, cars could change from one state to another. If a car arrives
and another car being washed, the arrival car becomes a waiting car in
queue and if there is at least one car waiting in queue the car wash
facility is available, then a waiting car becomes a being washed state.

Source code (Form Java, Java, JDK 8.0)
 Carwashmodel
o Main.form
o Main.java
o Class
 MultiServerMultiQueueSystem.java
 MultiServerSingleQueueSystem.java
 RandomGeneratorHelper.java
 SingleJobServerSystem.java
 SingleServerQueueSystem.java